# increase take performance by directly writing to buffer