--- a/py-polars/Cargo.toml
+++ b/py-polars/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "py-polars"
-version = "0.0.21"
+version = "0.0.22"
 authors = ["ritchie46 <ritchie46@gmail.com>"]
 edition = "2018"