@Dandandan I benchmarked the joins with a take kernel that directly writes to a buffer. Some thoughts:

* Still had to do quite some low-level memory management regarding the capacity of the buffer. Has `MutableBuffer` a utility similar to `Vec::push` that resizes when needed?
* Don't know the required size beforehand. I choose 1.5 times the expected size.
* I think a buffer should be shrunk before freezing?

The benchmarks in #192 improved from:

```
470 ms
477 ms
489 ms
```

to

```
338 ms
338 ms
342 ms
```

I saw that `Arrow::Stringarray::from_vec` also has the `Vec/memcpy` approach so assume this is widely used Arrow. I don't know where we could have this discussion in Arrow, but I'd like some centralized logic for the buffer relocation, instead of checking capacity/ reserving in every kernel where we like to use MutableBuffers directly.