@Dandandan 2020-11-30
Interesting! Good to see that this further speeds up things.

`MutableBuffer` indeed resizes when needed simarly to `Vec`, I think it does some small things differently related to cache alignment but it also keeps track of remaining capacity in the same way.

> Don't know the required size beforehand. I choose 1.5 times the expected size.

I wonder whether it would be faster/better to do two iterations, one to calculate the offsets and then use the `length_so_far` to allocate the string buffer and copy the strings. At least you don't need to over-allocate / reallocate if you underestimated and tune that factor.
Sometimes two loops can also be even faster than one that needs to fetch more data.