@Dandandan 2020-11-30
> I saw that Arrow::Stringarray::from_vec also has the Vec/memcpy approach so assume this is widely used Arrow. I don't know where we could have this discussion in Arrow, but I'd like some centralized logic for the buffer relocation, instead of checking capacity/ reserving in every kernel where we like to use MutableBuffers directly.

I also think that could be useful, at least to have something a `MutableBuffer<T>::new(capacity: usize)`
which uses `std::mem::sizeof::<T> * capacity` instead of allocating the `capacity` nr of bytes. That would make it an easier API to work with.