@ritchie46 2020-11-30
> MutableBuffer indeed resizes when needed simarly to Vec, I think it does some small things differently related to cache alignment but it also keeps track of remaining capacity in the same way.

Which method do you need to call for that? If I use the trait `io::Write::write` it panics when it exceeds the capacity. Does there exist a method that does implicit allocation similar to `Vec::push`?

> I wonder whether it would be faster/better to do two iterations, one to calculate the offsets and then use the length_so_far to allocate the string buffer and copy the strings. At least you don't need to over-allocate / reallocate if you underestimated and tune that factor.

That sounds interesting to find out. When you exceed the capacity you are basically back to `memcpy`.

>  Sometimes two loops can also be even faster than one that needs to fetch more data.

I guess it can be more cache optimal or simpler for (auto)vectorization. :thinking: 

> I also think that could be useful, at least to have something a MutableBuffer<T>::new(capacity: usize)
which uses std::mem::sizeof::<T> * capacity instead of allocating the capacity nr of bytes. That would make it an easier API to work with (similar to Vec<T>) 

Yes.. And then a `MutableBuffer::push::<T>()` with implicit alloc. :rocket: