@Dandandan 2020-11-30
Yeah mainly more cache optimal but can also be better branch prediction / prefetching or better compiler optimizations indeed. 

I recently watched this great example in this talk: https://youtu.be/o54YMz8zvCY?t=858