@ritchie46 2020-11-30
That was an interesting talk and insight. It's indeed faster! 

```
253 ms
250 ms
254 ms
```